Extract `UctBundle`
